feat(suite): added reusable hook for total portfolio balance
